Mejora al error sql mostrado en el cli, mejora al instancia regenerar

+++ php/consola/comandos/comando_instancia.php
@@ -214,8 +214,10 @@ class comando_instancia extends comando_toba
 		$this->consola->enter();
 		$i = $this->get_instancia();
 		$this->consola->lista( $i->get_parametros_db(), 'BASE' );
-		if ( $this->consola->dialogo_simple('Desea eliminar los datos de la INSTANCIA?') ) {
-			$i->eliminar_base();
+		if ($i->existe_modelo()) {
+			if ( $this->consola->dialogo_simple('Desea eliminar los datos de la INSTANCIA?') ) {
+				$i->eliminar_base();
+			}
 		}
 		$this->get_instancia()->cargar();
 	}
